Allow user to define a branch to rebase first
When given a list of branches to rebase, it will normally go thourgh
them in aplhabetical order.
This code adds a config (first_rebase_branch) or ENV option
(GIT_UP_FIRST_REBASE_BRANCH), that allows the user to define which
branch to rebase first.

e.g. This can be handy when rebasing one of the branches fails before
master, suppling 'master' here would mean that it is rebased, allowing a
manual rebase later.